Refactor: Extract class ExcludeFileFilter
